Separate each machine in undoMachine into files
