+++ visualizer/src/Project/service/undoMachine.js
@@ -1,157 +1,9 @@
-import { actions, assign, forwardTo, Machine, send, sendParent, spawn } from 'xstate';
+import { actions, assign, forwardTo, Machine, spawn } from 'xstate';
 import { fromEventBus } from './eventBus';
+import createHistoryMachine from './undo/historyMachine';
+import createLabelHistoryMachine from './undo/labelHistoryMachine';
 
 const { pure } = actions;
-
-/** Records a stack of snapshots of an actor's state before each action.
- * Sends SAVE events to an actor and stores the actor's responses to send back when undoing or redoing actions.
- */
-const createHistoryMachine = (actor) =>
-  Machine(
-    {
-      context: {
-        actor,
-        past: [],
-        future: [],
-      },
-      initial: 'idle',
-      states: {
-        idle: {
-          on: {
-            EDIT: 'saving',
-            UNDO: 'restoringPast',
-            REDO: 'restoringFuture',
-          },
-        },
-        saving: {
-          entry: 'getSnapshot',
-          on: {
-            RESTORE: { target: 'idle', actions: 'saveSnapshotInPast' },
-          },
-          exit: sendParent('SAVED'),
-        },
-        restoringPast: {
-          entry: 'restorePast',
-          exit: 'movePastToFuture',
-          on: {
-            RESTORED: { target: 'idle', actions: 'forwardToParent' },
-          },
-          // fallback in case actor does not respond
-          after: {
-            500: {
-              target: 'idle',
-              actions: [
-                sendParent('RESTORED'),
-                // TODO: log what actors are not restoring
-                () => console.log('could not restore actor'),
-              ],
-            },
-          },
-        },
-        restoringFuture: {
-          entry: 'restoreFuture',
-          exit: 'moveFutureToPast',
-          on: {
-            RESTORED: { target: 'idle', actions: 'forwardToParent' },
-          },
-          // fallback in case actor does not respond
-          after: {
-            500: {
-              target: 'idle',
-              actions: [
-                sendParent('RESTORED'),
-                // TODO: log what actors are not restoring
-                () => console.log('could not restore actor'),
-              ],
-            },
-          },
-        },
-      },
-    },
-    {
-      actions: {
-        forwardToParent: sendParent((context, event) => event),
-        getSnapshot: send('SAVE', { to: (context) => context.actor }),
-        saveSnapshotInPast: assign((context, event) => ({
-          past: [...context.past, event],
-          future: [],
-        })),
-        restorePast: send((context) => context.past[context.past.length - 1], {
-          to: (context) => context.actor,
-        }),
-        restoreFuture: send((context) => context.future[context.future.length - 1], {
-          to: (context) => context.actor,
-        }),
-        movePastToFuture: assign({
-          past: (context) => context.past.slice(0, context.past.length - 1),
-          future: (context) => [...context.future, context.past[context.past.length - 1]],
-        }),
-        moveFutureToPast: assign({
-          past: (context) => [...context.past, context.future[context.future.length - 1]],
-          future: (context) => context.future.slice(0, context.future.length - 1),
-        }),
-      },
-    }
-  );
-
-/** Records edited labels before and after each action. */
-function createLabelHistoryMachine(actor) {
-  return Machine(
-    {
-      id: 'label-history', // TODO: add actor to ID
-      context: {
-        actor,
-        // TODO: switch from snapshot list to object with action IDs as keys
-        // not every action edits all labels, so use ID to check if there's a snapshot for the undone/redone action
-        past: [],
-        future: [],
-      },
-      entry: send({ type: 'HISTORY_REF' }, { to: actor }),
-      initial: 'idle',
-      states: {
-        idle: {
-          on: {
-            SAVE_LABELS: { actions: 'addSnapshotToPast' },
-            UNDO: 'undoing',
-            REDO: 'redoing',
-          },
-        },
-        undoing: {
-          entry: 'undo',
-          always: 'idle',
-          exit: 'movePastToFuture',
-        },
-        redoing: {
-          entry: 'redo',
-          always: 'idle',
-          exit: 'moveFutureToPast',
-        },
-      },
-    },
-    {
-      actions: {
-        addSnapshotToPast: assign({
-          past: (ctx, evt) => [...ctx.past, [evt.initialLabels, evt.editedLabels]],
-          future: [],
-        }),
-        undo: send((ctx, evt) => ({ type: 'EDITED', ...ctx.past[ctx.past.length - 1][0] }), {
-          to: (ctx) => ctx.actor,
-        }),
-        redo: send((ctx, evt) => ({ type: 'EDITED', ...ctx.future[ctx.future.length - 1][1] }), {
-          to: (ctx) => ctx.actor,
-        }),
-        movePastToFuture: assign({
-          past: (context) => context.past.slice(0, context.past.length - 1),
-          future: (context) => [...context.future, context.past[context.past.length - 1]],
-        }),
-        moveFutureToPast: assign({
-          past: (context) => [...context.past, context.future[context.future.length - 1]],
-          future: (context) => context.future.slice(0, context.future.length - 1),
-        }),
-      },
-    }
-  );
-}
 
 const createUndoMachine = ({ eventBuses }) =>
   Machine(
